requirements.txt
astroid==2.8.4
black==21.9b0
CacheControl==0.12.7
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
firebase-admin==5.0.3
Flask==2.0.2
Flask-Login==0.5.0
Flask-WTF==0.15.1
google-api-core==2.2.1
google-api-python-client==2.28.0
google-auth==2.3.2
google-auth-httplib2==0.1.0
google-cloud-core==2.1.0
google-cloud-datastore==2.3.0
google-cloud-firestore==2.3.4
google-cloud-storage==1.42.3
google-crc32c==1.3.0
google-resumable-media==2.1.0
googleapis-common-protos==1.53.0
grpcio==1.41.1
grpcio-status==1.41.1
httplib2==0.20.1
idna==3.3
isort==5.10.0
itsdangerous==2.0.1
Jinja2==3.0.2
lazy-object-proxy==1.6.0
libcst==0.3.21
MarkupSafe==2.0.1
mccabe==0.6.1
msgpack==1.0.2
mypy-extensions==0.4.3
packaging==21.2
pathspec==0.9.0
platformdirs==2.4.0
proto-plus==1.19.7
protobuf==3.19.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pylint==2.11.1
pyparsing==2.4.7
python-dotenv==0.19.1
PyYAML==6.0
regex==2021.10.23
requests==2.26.0
rsa==4.7.2
six==1.16.0
spotipy==2.19.0
toml==0.10.2
tomli==1.2.2
typing-extensions==3.10.0.2
typing-inspect==0.7.1
uritemplate==4.1.1
urllib3==1.26.7
Werkzeug==2.0.2
wrapt==1.13.3
WTForms==2.3.3
